refactor(api): move api/admin/organizations/{id} to src/team
